requirements.txt
aiofiles
annotated-types
anyio
certifi
cffi
contourpy
cycler
eval_type_backport
fonttools
h11
httpcore
httpx
hume
idna
kiwisolver
matplotlib
numpy
opencv-python
packaging
pandas
pillow
pycparser
pydantic
pydantic_core
pydub
pyparsing
python-dateutil
python-dotenv>=1.0.0
six
sniffio
sklearn
sounddevice
typing_extensions
websockets
flask
transformers
flask_cors
gunicorn
torch
supabase
nltk
openai
colorama